move binned data simulator to tests/helpers
