+++ tests/helpers.py
@@ -1,0 +1,35 @@
+import numpy as np
+import xarray as xr
+
+def simulate_binned_data(shape: tuple, dims: list):
+    """Generate a fake xr.DataArray as those generated by binning
+
+    Used for testing purpouses
+
+    Args:
+        shape: Shape ouf the data
+        dims: name of the dimensions
+
+    Returns:
+        the simulated data array
+    """
+    assert len(dims) == len(
+        shape,
+    ), "number of dimesions and data shape must coincide"
+
+    return xr.DataArray(
+        data=np.random.rand(*shape),
+        coords={d: np.linspace(-1, 1, s) for d, s in zip(dims, shape)},
+        attrs={
+            "metadata": {
+                "list": [1, 2, 3],
+                "string": "asdf",
+                "int": 1,
+                "float": 1.0,
+                "bool": True,
+                "nested": {
+                    "nestedentry": 1.0,
+                },
+            },
+        },
+    )
